refactor: update project
