+++ src/main/java/com/pe/relari/employee/model/api/EmployeeRequest.java
@@ -53,13 +53,13 @@ public class EmployeeRequest {
 
     @Schema(
             description = "Sexo del Empleado",
-            name = "sex",
+            name = "gender",
             type = "String",
             example = "M o F",
             required = true)
     @NotBlank
     @Pattern(regexp = "[M|F]")
-    private String sex;
+    private String gender;
 
     @Schema(
             description = "Cargo del Empleado",
